add authorization logics deleting a users task status

+++ src/tasks/tasks.service.ts
@@ -48,9 +48,12 @@ export class TasksService {
 
   
 
-    async deleteTaskById(id: number):Promise<void> {
+    async deleteTaskById(
+        id: number,
+        user: User
+        ):Promise<void> {
         
-    const result = await this.taskRepository.delete(id)
+    const result = await this.taskRepository.delete({id, userId: user.id})
 
     if(result.affected === 0){
         throw new NotFoundException(`Task with id ${id} not found`)
